requirements.txt
mcp>=0.9.0
anthropic>=0.18.0
openai>=1.0.0
tiktoken>=0.5.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
redis>=5.0.0
python-dotenv>=1.0.0
spacy>=3.7.0
asyncio
faker>=20.0.0
gradio>=4.0.0
psutil>=5.9.0
httpx>=0.24.0
pyyaml>=6.0
requests>=2.31.0
